pyproject.toml
[tool.poetry]
name = "deepdock"
version = "0.1.0"
description = ""
authors = ["Rico Meinl"]

[tool.poetry.dependencies]
python = ">=3.7.1"
numpy = "1.21.0"
pandas = "^1.3.5"
matplotlib = "^3.5.1"
plyfile = "^0.7.4"
networkx = "^2.6.3"
scikit-learn = "^1.0.2"
rdkit-pypi = "^2021.9.4"
torch = {url = "https://download.pytorch.org/whl/cu111/torch-1.9.0%2Bcu111-cp37-cp37m-linux_x86_64.whl"}
torch-scatter = {url = "https://data.pyg.org/whl/torch-1.9.0%2Bcu111/torch_scatter-2.0.9-cp37-cp37m-linux_x86_64.whl"}
torch-sparse = {url = "https://data.pyg.org/whl/torch-1.9.0%2Bcu111/torch_sparse-0.6.12-cp37-cp37m-linux_x86_64.whl"}
torch-cluster = {url = "https://data.pyg.org/whl/torch-1.9.0%2Bcu111/torch_cluster-1.5.9-cp37-cp37m-linux_x86_64.whl"}
torch-spline-conv = {url = "https://data.pyg.org/whl/torch-1.9.0%2Bcu111/torch_spline_conv-1.2.1-cp37-cp37m-linux_x86_64.whl"}
torch-geometric = "^2.0.3"
pymesh = "^1.0.2"
click = "^8.0.3"
biopython = "^1.79"

[tool.poetry.dev-dependencies]
ipykernel = "^6.7.0"

[tool.poetry.scripts]
prep_target = "deepdock.prep_target:main"

[build-system]
requires = ["setuptools", "poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
